Move merge_service_config.go over to agent/consul/configentry
