+++ agent/configentry/merge_service_config_test.go
@@ -1,4 +1,4 @@
-package consul
+package configentry
 
 import (
 	"testing"
